Use the right blocks for the document outline
