+++ packages/editor/src/components/document-outline/style.scss
@@ -81,3 +81,15 @@
 .document-outline__item-content {
 	padding: 1px 0;
 }
+
+.editor-document-outline.has-no-headings {
+	& > svg {
+		margin-top: $grid-unit-30 + $grid-unit-05;
+	}
+	& > p {
+		padding-left: $grid-unit-40;
+		padding-right: $grid-unit-40;
+	}
+	text-align: center;
+	color: $gray-700;
+}
